Setup alert words.
Notifies when a word from a list appears in subtitles.

A few things to note:

* To take input from the extension, we use chips provided
by materializecss. One quirk of chips is that the data words
are expected to be in the format {tag: 'word'}. For the ease
of moving these words around, I added helper functions to deal
with that format.

* background.js 's sole purpose is for notifications as of now.

* subtitle checks are case insensitive. A notification is sent for
every word found from the alertWords array. We might want to play
around with this to ensure no enormous spam.

Few links for further reference:
notifications api:
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/notifications
We use notifications.create().

about background scripts:
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Anatomy_of_a_WebExtension#Background_scripts

chips:
https://materializecss.com/chips.html