Fix not notifying alert words.
The issue was, when there were no captions, the function call
document.querySelector(selector) in getSubtitlesContent was
returning null to which we were doing .textContent. This
failed without showing an error.

Also refactered a bit. Made it use setTimeout to have
variable wait times based on whether a word appeared.
And changed the test about previous caption to check
if current caption starts with previous caption. This
reduces a lot of notifications spam.

We should soon clean up the code around here to use
a single word, subtitles or captions for consistency.